major refactor to flash code
