+++ sed/loader/fel/__init__.py
@@ -1,0 +1,14 @@
+"""sed.loader.fel module easy access APIs
+
+"""
+from .buffer import BufferFileHandler
+from .dataframe import DataFrameCreator
+from .multiindex import MultiIndexCreator
+from .parquet import ParquetHandler
+
+__all__ = [
+    "BufferFileHandler",
+    "DataFrameCreator",
+    "MultiIndexCreator",
+    "ParquetHandler",
+]
